extract stream writer into a separate file
