+++ ydb/src/client_topic/topicwriter/mod.rs
@@ -1,6 +1,8 @@
+pub mod connection;
 pub mod message;
 pub mod message_write_status;
 pub mod partitioning;
+pub mod stream_writer;
 pub mod writer;
 pub mod writer_options;
 pub mod writer_reception_queue;
